# nodejsss/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "College Project API",
    "version": "1.0.0",
    "description": "The admin login system is implemented using secure authentication and role-based authorization with JSON Web Token (JWT). When a user attempts to log in, the system verifies the email, checks whether the user has an admin role, and compares the entered password with the stored hashed password. If the credentials are valid, a JWT is generated containing the user’s identity and role information and is sent to the client. The client stores this token in local storage and includes it in the authorization header for subsequent API requests. Middleware is used on the server to intercept incoming requests, verify the JWT, and ensure that the user has admin privileges before granting access to protected routes. This ensures that only authenticated administrators can access sensitive functionalities, providing a secure, scalable, and stateless authentication mechanism."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/admin/login": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin Login",
        "description": "Authenticate an admin and return a JWT token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "admin@college.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "admin123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "admin": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/admin/profile": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get Admin Profile",
        "description": "Retrieve the profile data of the currently authenticated admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (Admin only)"
          },
          "404": {
            "description": "Admin not found"
          }
        }
      }
    },
    "/api/students": {
      "get": {
        "tags": [
          "Students"
        ],
        "summary": "Get all students",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of students"
          }
        }
      },
      "post": {
        "tags": [
          "Students"
        ],
        "summary": "Create new student",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Student created"
          }
        }
      }
    },
    "/api/students/{id}": {
      "get": {
        "tags": [
          "Students"
        ],
        "summary": "Get student by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Student details"
          }
        }
      },
      "put": {
        "tags": [
          "Students"
        ],
        "summary": "Update student",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Students"
        ],
        "summary": "Delete student",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Student deleted"
          }
        }
      }
    }
  }
}